add check for rmdir

+++ packages/openapi-generator/src/generator.ts
@@ -1,6 +1,6 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
 
-import { promises as promisesFs } from 'fs';
+import { promises as promisesFs,existsSync } from 'fs';
 import { resolve, parse } from 'path';
 import {
   createLogger,
@@ -38,12 +38,17 @@ export async function generate(options: GeneratorOptions): Promise<void[]> {
     resolve(options.input.toString(), 'service-mapping.json');
 
   if (options.clearOutputDir) {
-    await rmdir(options.outputDir, { recursive: true });
+    if(existsSync(options.outputDir)) {
+      await rmdir(options.outputDir, { recursive: true });
+    }
   }
 
   const vdmMapping = readServiceMapping(options);
   const uniqueNameGenerator = new UniqueNameGenerator('-');
   const inputFilePaths = await getInputFilePaths(options.input);
+
+  console.log('Generation for the following file paths:' + JSON.stringify(inputFilePaths))
+
   const promises = inputFilePaths.map(async inputFilePath => {
     const uniqueServiceName = uniqueNameGenerator.generateAndSaveUniqueName(
       parseServiceName(inputFilePath)
@@ -198,8 +203,11 @@ async function generateFromFile(
     JSON.stringify(openApiDocument, null, 2)
   );
 
+  console.log('Dummy part of  generation finished for' + serviceName)
   await generateOpenApiService(convertedInputFilePath, serviceDir);
+  console.log('TS client part of generation finished for' + serviceName)
   await generateSDKSources(serviceDir, parsedOpenApiDocument, options);
+  console.log('JS client part of generation finished for' + serviceName)
 }
 
 /**
